refactor(errorhanlder): organise error handler
re 37
