+++ src/core/usecase/ErrorHandler/IErrorHandler.ts
